Hoefully fix float midi np (should be int)

+++ train/train.py
@@ -336,7 +336,10 @@ def train():
                 audio_input = audio.reshape((1, -1, s.sample_rate))
                 pred_midi, pred_len, pred_tpb = model(torch.Tensor(audio_input))
                 out = Song.from_np(
-                    pred_midi[0], None, int(pred_tpb[0]), int(pred_len[0])
+                    pred_midi[0].astype(np.uint16),
+                    None,
+                    int(pred_tpb[0]),
+                    int(pred_len[0]),
                 )
 
                 wandb.log(
